+++ tests/sentry/api/endpoints/test_organization_metric_data.py
@@ -462,7 +462,7 @@ class OrganizationMetricDataTest(MetricsAPIBaseTestCase):
         assert response.status_code == 400
         assert (
             response.data["detail"]
-            == 'Failed to parse query: Query !release:"0.99.0 (*)" contains wildcards'
+            == "Failed to parse query: Release Health Queries don't support wildcards"
         )
 
     def test_pagination_offset_without_orderby(self):
